ssss

+++ src/main/java/fpt/aptech/eventsphere/repositories/UserDetailsRepository.java
@@ -7,11 +7,12 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
 import java.util.Optional;
 import java.util.List;
 
 @Repository
-public interface UserDetailsRepository extends JpaRepository<UserDetails, Integer> {
+public interface UserDetailsRepository extends JpaRepository<UserDetails, BigDecimal> {
 
     Optional<UserDetails> findByUser(Users user);
